PautaController incluye validación de existencia de ICI Agrega boton "eliminar ICI" a ejecutivo-evaluaciones.blade.php (base ejecutivo) Se agrega ruta para eliminar ici

+++ app/Http/Controllers/PautaController.php
@@ -9,19 +9,21 @@ use Illuminate\Http\Request;
 class PautaController extends Controller
 {
     public function resetici($evaluacionid){
-        Respuesta::where('evaluacion_id', $evaluacionid)->where('origen_id',1)->delete();
         $evaluacion = Evaluacion::find($evaluacionid);
-        $respuestas = Respuesta::where('evaluacion_id', $evaluacionid)->where('origen_id',2)->get();
-        foreach ($respuestas as $respuesta) {
-            $respuesta->origen_id = 1;
-            $respuesta->save();
+        $tieneICI = $evaluacion->respuestas->max('origen_id') == 2;
+        if ($tieneICI) {
+            Respuesta::where('evaluacion_id', $evaluacionid)->where('origen_id',1)->delete();
+            $respuestas = Respuesta::where('evaluacion_id', $evaluacionid)->where('origen_id',2)->get();
+            foreach ($respuestas as $respuesta) {
+                $respuesta->origen_id = 1;
+                $respuesta->save();
+            }
+
+            $evaluacion->ici = null;
+            $evaluacion->user_ici = null;
+            $evaluacion->fecha_ici = null;
+            $evaluacion->save();
         }
-
-        $evaluacion->ici = null;
-        $evaluacion->user_ici = null;
-        $evaluacion->fecha_ici = null;
-        $evaluacion->save();
-
         return redirect(route('evaluacions.index', [$evaluacionid]))->with("status", "Se borro ICI correctamente");
 
     }
